General: - Sound finally works! ^-^ - Code cleanup and refactoring.
Board:
- added mapper 29 (homebrew mapper for Glider);
- added mapper 50;
- added mapper 61;
- added mapper 62;
- added mapper 91;
- added mapper 148;
- added mapper 149;
- added mapper 172;
- added mapper 173;
- added mapper 188;
- added mapper 216.

+++ src/nesimulare/core/audio/MMC5SoundChip.java
@@ -81,10 +81,10 @@ public class MMC5SoundChip implements ExpansionSoundChip {
     }
     
     @Override
-    public short mix() {
-        short output = (short)pulse1.getOutput();
-        output += (short)pulse2.getOutput();
-        output += (short)pcm.getOutput();
+    public int mix() {
+        int output = pulse1.getOutput();
+        output += pulse2.getOutput();
+        output += pcm.getOutput();
         
         return output;
     }
